backend/requirements.txt
abstract_singleton==1.0.1
africastalking==1.2.7
aiofiles==0.6.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.5.0
anyio==4.0.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.2.1
autopep8==2.0.4
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
blis==0.7.10
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colored==2.2.3
command-not-found==0.3
confection==0.1.3
contextlib2==21.6.0
contourpy==1.1.1
coverage==7.3.1
cryptography==41.0.3
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.8
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.7
distro==1.8.0
distro-info==1.0
Django==4.2.6
django-crispy-forms==2.1
djangorestframework==3.15.1
dnspython==2.4.2
docker==6.1.3
duckduckgo-search==3.8.5
emoji==2.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.3
execnet==2.0.2
executing==1.2.0
fabric==3.2.2
Fabric3==1.14.post1
filelock==3.12.4
flake8==6.1.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.42.1
frozenlist==1.4.0
fs==2.4.16
gast==0.5.4
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.36
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
greenlet==2.0.2
gTTS==2.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
identify==2.5.29
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
invoke==2.2.0
ipython==8.15.0
isort==5.12.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==24.2.0
kiwisolver==1.4.5
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
loguru==0.7.2
louis==3.20.0
lxml==4.9.3
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mistune==3.0.1
mkdocs==1.5.2
mongoengine==0.27.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
murmurhash==1.0.10
mypy==1.5.1
mypy-extensions==1.0.0
netifaces==0.11.0
nodeenv==1.8.0
numpy==1.26.0
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
openai==0.28.0
openapi-python-client==0.15.2
orjson==3.9.7
otree==5.10.3
outcome==1.2.0
overprint==0.0.2
packaging==23.1
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pinecone-client==2.2.4
pipenv==2023.9.8
pkgtools==0.7.3
platformdirs==3.10.0
playsound==1.3.0
plotly==5.24.1
pluggy==1.3.0
pocketsphinx==5.0.3
pre-commit==3.4.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.24.0
pycodestyle==2.11.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.12
pydantic_core==2.6.3
pydub==0.25.1
pyflakes==3.1.0
pygame==2.5.2
PyGithub==1.59.1
Pygments==2.16.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.8.0
pymacaroons==0.13.0
pymdown-extensions==10.3
pymongo==4.5.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
pyTelegramBotAPI==4.18.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-integration==0.2.3
pytest-mock==3.11.1
pytest-recording==0.13.0
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==1.0.0
python-multipart==0.0.5
python-telegram-bot==21.3
pythran==0.14.0
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
pyyaml_env_tag==0.1
qrcode==7.4.2
readability-lxml==0.8.1
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
schema==0.7.5
scipy==1.11.2
SecretStorage==3.3.3
selenium==4.12.0
shellingham==1.5.3
simplejson==3.19.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.13.0
SQLAlchemy==1.3.22
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.6.2
starlette==0.14.1
sympy==1.12
systemd-python==234
tenacity==9.0.0
termcolor==2.3.0
thinc==8.1.10
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
traitlets==5.1.1
TravisPy==0.3.5
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.7.0
typing_extensions==4.8.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.13.4
vcrpy==4.2.1
virtualenv==20.24.5
virtualenv-clone==0.5.7
wadllib==1.3.6
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.5
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.1
Werkzeug==3.0.3
withlog==0.0.3
wrapt==1.15.0
wsproto==1.2.0
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
